Side 14 Chuong 4

+++ BaiThucHanh/model/Nhanvien.java
@@ -2,7 +2,7 @@ package model;
 
 public class Nhanvien {
     protected String ten;
-    protected long luong;
+    protected float luong;
     public Nhanvien(){
 
     }
